agregando cambio a todos el tipo de variable de la id

+++ acad-antecedenteMedico-service/src/main/java/com/acad/antecedentemedicoservice/Entity/AntecedenteMedico.java
@@ -17,7 +17,7 @@ public class AntecedenteMedico {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idAntecedenteMedico")
-    private Integer idAntecedenteMedico;
+    private Long idAntecedenteMedico;
 
     @Column(name = "edad", length = 45)
     private String edad;
